requirements.txt
Flask
flask-cors
python-dotenv
Werkzeug
google-genai
Pillow
openai
deep-translator
gTTS
gunicorn
google-generativeai